+++ agent/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2024 Inria
+
+"""Default agent for Upkie wheeled bipeds."""
+
+from .whole_body_controller import WholeBodyController
+
+__all__ = [
+    "WholeBodyController",
+]