+++ agent/sagittal_balance/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2024 Inria
+
+from .mpc_balancer import MPCBalancer
+from .sagittal_balancer import SagittalBalancer
+
+__all__ = [
+    "MPCBalancer",
+    "SagittalBalancer",
+]