+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- agent: Import default agent from the Pink balancer
 - Development workflow based on Pixi
 - control: Create `upkie.control` submodule
 - control: Import MPC balancer class from the Pink balancer